+++ packages/asc-ui/src/components/Alert/index.ts
@@ -1,0 +1,2 @@
+export { default as AlertStyle } from './AlertStyle'
+export { default as AlertMessage } from './Alert'